[REFACTOR] notibell의 fetchNotificationData를 hook으로 분리
